Razbijanje velikih fajlova u bank servisu
