+++ internal/bank/company_test.go
