+++ internal/bank/recipient_repo.go
@@ -1,0 +1,1 @@
+package bank
